requirements.txt
torch==2.0.0
torchsummary==1.5.1
torchvision==0.15.1
opencv-python==4.7.0.72
numpy==1.23.5
matplotlib==3.7.0
Pillow==9.4.0